- [CLN] Logic moved out of inits
