+++ examples/getting_started/getting_started.py
@@ -33,6 +33,11 @@ import pooch
 # %%
 
 # Pulling data example
+import subsurface.interfaces.to_liquid_earth
+import subsurface.io.profiles.profiles_core
+import subsurface.io.topography.topo_core
+import subsurface.io.wells.wells_interface
+
 model_file = pooch.retrieve(
     url="https://github.com/cgre-aachen/gempy_data/raw/master/"
         "data/subsurface/example1.zip",
@@ -116,7 +121,7 @@ ss.visualization.pv_plot(
 # %%
 
 # UnstructuredData
-borehole_location_struct = ss.io.borehole_location_to_unstruct(
+borehole_location_struct = subsurface.io.wells.wells_interface.borehole_location_to_unstruct(
     collar_file=data_path + '/wells.csv',
     read_collar_kwargs={
         'usecols': ['Index', 'X', 'Y', 'Altitude'],
@@ -153,7 +158,7 @@ ss.visualization.pv_plot(
 # %%
 
 # StructuredData
-topo_structured_data = ss.io.read_structured_topography(data_path + '/DEM50.tif')
+topo_structured_data = subsurface.io.topography.topo_core.read_structured_topography(data_path + '/DEM50.tif')
 topo_structured_data
 # %%
 # Remove outliers
@@ -181,13 +186,13 @@ ss.visualization.pv_plot(
 
 # %%
 
-profiles_traces = ss.io.profiles.lineset_from_trace(
+profiles_traces = subsurface.io.profiles.profiles_core.lineset_from_trace(
     data_path + '/Profiles_cropped/Profile_PyVista.shp',
     idx=range(13)
 )
 
 # %%
-profiles_trisurf_list, profiles_mesh_list = ss.io.profiles.create_tri_surf_from_traces_texture(
+profiles_trisurf_list, profiles_mesh_list = subsurface.io.profiles.profiles_core.create_tri_surf_from_traces_texture(
     data_path + '/Profiles_cropped/Profile_PyVista.shp',
     path_to_texture=[
         data_path + '/Profiles_cropped/profile001.png',
@@ -247,19 +252,19 @@ ss.visualization.pv_plot(
 # ----------------
 
 # %%
-ss.interfaces.base_structs_to_binary_file(data_path + '/gempy_base',
-                                          gempy_unstructured_data)
-ss.interfaces.base_structs_to_binary_file(data_path + '/wells',
-                                          wells_unstructured_data)
-ss.interfaces.base_structs_to_binary_file(data_path + '/topo',
-                                          topo_structured_data)
-ss.interfaces.base_structs_to_binary_file(data_path + '/collars',
-                                          borehole_location_struct)
+subsurface.interfaces.to_binary.base_structs_to_binary_file(data_path + '/gempy_base',
+                                                            gempy_unstructured_data)
+subsurface.interfaces.to_binary.base_structs_to_binary_file(data_path + '/wells',
+                                                            wells_unstructured_data)
+subsurface.interfaces.to_binary.base_structs_to_binary_file(data_path + '/topo',
+                                                            topo_structured_data)
+subsurface.interfaces.to_binary.base_structs_to_binary_file(data_path + '/collars',
+                                                            borehole_location_struct)
 
 for e, tri_surf in enumerate(profiles_trisurf_list):
-    ss.interfaces.base_structs_to_binary_file(data_path + f'/profile_{e}_mesh',
-                                              tri_surf.mesh)
-    ss.interfaces.base_structs_to_binary_file(
+    subsurface.interfaces.to_binary.base_structs_to_binary_file(data_path + f'/profile_{e}_mesh',
+                                                                tri_surf.mesh)
+    subsurface.interfaces.to_binary.base_structs_to_binary_file(
         data_path + f'/profile_{e}_texture_C',
         tri_surf.texture,
         order='C')
