+++ subsurface/io/__init__.py
@@ -2,8 +2,10 @@ import xarray as xr
 
 from .profiles import *
 from .wells.welly_reader import *
-from .topography import read_structured_topography, read_unstructured_topography
-from .wells import borehole_location_to_unstruct, read_wells_to_unstruct
+from .topography.topo_core import read_structured_topography, \
+    read_unstructured_topography
+from .wells import read_wells_to_unstruct
+from .wells.wells_interface import borehole_location_to_unstruct
 
 
 def read_unstruct(path, **kwargs):
