+++ subsurface/utils/__init__.py
@@ -1,6 +1,3 @@
-from pathlib import Path
 
 
-def get_extension(path):
-    p = Path(path)
-    return p.suffix
+
