+++ subsurface/utils/utils_core.py
@@ -1,0 +1,6 @@
+from pathlib import Path
+
+
+def get_extension(path):
+    p = Path(path)
+    return p.suffix
